shellfoundry was improved to be used directly from command line:
https://gehrcke.de/2014/02/distributing-a-python-command-line-application/
